requirements.txt
amqp==5.1.1
api==0.0.7
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
awsebcli==3.20.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.24.9
botocore==1.27.9
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.3
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0
django-celery-beat==2.3.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-redis==5.2.0
django-storages==1.12.3
django-timezone-field==5.0
djangorestframework==3.12.0
djangorestframework-simplejwt==4.6.0
environ==1.0
fcm-django==1.0.12
firebase-admin==5.2.0
future==0.16.0
geographiclib==1.52
geolocator==0.1.1
geopy==2.2.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
jmespath==1.0.0
kombu==5.2.4
lxml==4.8.0
msgpack==1.0.4
nose==1.3.7
numpy==1.22.3
outcome==1.2.0
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
pdfminer.six==20220319
pdfplumber==0.6.2
pikepdf==5.1.2
Pillow==9.1.0
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.20.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.3.0
pyngrok==5.1.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.3
python-dotenv==0.19.2
pytz==2022.1
pywin32==304
PyYAML==5.4.1
razorpay==1.3.0
redis==4.3.4
requests==2.27.1
rsa==4.8
s3transfer==0.6.0
selenium==4.4.3
semantic-version==2.8.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tabulate==0.8.10
termcolor==1.1.0
trio==0.21.0
trio-websocket==0.9.2
twilio==7.8.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
Wand==0.6.7
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.2.0
